refactor: move common binance query functions to utils
